Enum zng_view_api::access::AccessRole

#[non_exhaustive]
pub enum AccessRole {

    Button,
    CheckBox,
    GridCell,
    Link,
    MenuItem,
    MenuItemCheckBox,
    MenuItemRadio,
    Option,
    ProgressBar,
    Radio,
    ScrollBar,
    SearchBox,
    Slider,
    SpinButton,
    Switch,
    Tab,
    TabPanel,
    TextInput,
    TreeItem,
    ComboBox,
    Grid,
    ListBox,
    Menu,
    MenuBar,
    RadioGroup,
    TabList,
    Tree,
    TreeGrid,
    Application,
    Article,
    Cell,
    Column,
    ColumnHeader,
    Definition,
    Document,
    Feed,
    Figure,
    Group,
    Heading,
    Image,
    List,
    ListItem,
    Math,
    Note,
    Row,
    RowGroup,
    RowHeader,
    Separator,
    Table,
    Term,
    ToolBar,
    ToolTip,
    Banner,
    Complementary,
    ContentInfo,
    Form,
    Main,
    Navigation,
    Region,
    Search,
    Alert,
    Log,
    Marquee,
    Status,
    Timer,
    AlertDialog,
    Dialog,

}

Accessibility role of a node in the accessibility tree.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Button

Clickable widgets that trigger a response when activated by the user.

CheckBox

checkable interactive widget.

Must also set AccessState::Checked.

GridCell

Identifies a cell in a grid widget.

Link

Interactive reference to a resource

MenuItem

Indicates the widget is an option in a set of choices contained by a menu or menu-bar.

MenuItemCheckBox

Widget is a checkable option in a menu.

Must also set AccessState::Checked.

MenuItemRadio

Widget is a selectable option in a menu where only one option can be selected at a time.

Option

Selectable items in a list-box.

ProgressBar

Defines a widget that displays the progress status for tasks that take a long time.

The AccessState::Value and other value states define the progress.

Radio

Selectable items in a list where only one item may be selected at a time.

ScrollBar

Widget controls the scrolling of content within a viewing area.

Must also set AccessState::Controls and AccessState::Value to define the scroll widget and amount scrolled. By default the value min/max is 0/100.

SearchBox

Identifies a text-box that is used for searching.

Slider

Defines an input where the user selects a value from within a given range.

The AccessState::Value and other value states define the range and value.

SpinButton

Defines a type of range that expects the user to select a value from among discrete choices.

Switch

Identifies a check-box with named states.

Tab

Identifies a widget in a tab-list that selects the active tab in a tab-panel.

TabPanel

Identifies a container for the active tab.

TextInput

Identifies a widget that allows the input of free-form text.

TreeItem

Identifies an item in a tree widget.

ComboBox

Identifies a widget as an input that controls another widget, such as a list-box or grid, that can dynamically pop up to help the user set the value of that input.

Grid

Identifies a container of columns, rows and cells.

ListBox

Identifies a list of selectable items.

Menu

Identifies a composite widget that offers a list of choices to the user.

MenuBar

Identifies the part of a menu that always stays visible.

RadioGroup

Identifies a group of radio buttons.

TabList

Identifies the widget that serves as the container for a set of tabs. The selected tab content is shown in a TabPanel.

Tree

Widget that allows the user to select one or more items from a hierarchically organized collection.

TreeGrid

Identifies a widget as being grid whose rows can be expanded and collapsed in the same manner as for a tree.

Application

Indicates to assistive technologies that a widget and all of its children should be treated similar to a desktop application.

Article

Indicates a section of a page that could easily stand on its own.

Cell

Identifies a widget as being a cell in a tabular container that does not contain column or row header information.

Column

Identifies a column of cells within a tabular structure.

ColumnHeader

Identifies a widget as being a cell in a row contains header information for a column.

Definition

Indicates the widget is a definition of a term or concept.

Document

Focusable content within complex composite widgets or applications for which assistive technologies can switch reading context back to a reading mode.

Feed

Identifies a dynamic scrollable list of articles in which articles are added to or removed from either end of the list as the user scrolls.

Figure

Identify a figure inside page content where appropriate semantics do not already exist.

Group

Identifies a set of user interface objects that is not intended to be included in a page summary or table of contents by assistive technologies.

Heading

Defines a heading to a page or section, with AccessState::Level defining structure.

Image

Identifies a widget container that should be considered as a single image.

List

Identifies a list of items.

ListItem

Identifies an item inside a list of items.

Math

Indicates that the content represents a mathematical expression.

Note

Identifies a section whose content is parenthetic or ancillary to the main content.

Row

Identifies a row of cells within a tabular structure.

RowGroup

Identifies a group of rows within a tabular structure.

RowHeader

Identifies a cell containing header information for a row within a tabular structure.

Separator

Identifies a divider that separates and distinguishes sections of content or groups of menu items.

Table

Identifies the widget containing the role as having a non-interactive table structure containing data arranged in rows and columns.

Term

Identifies a word or phrase with an optional corresponding Definition.

ToolBar

Defines the containing widget as a collection of commonly used function buttons or controls represented in a compact visual form.

ToolTip

Identifies a contextual text bubble that displays a description for an element that appears on pointer hover or keyboard focus.

Banner

Identifies the global header, which usually includes a logo, company name, search feature, and possibly the global navigation or a slogan.

Complementary

Identifies a supporting section that relates to the main content.

ContentInfo

Identifies a footer, containing identifying information such as copyright information, navigation links, and privacy statements.

Form

Identify a group of widgets that are a register form.

Main

Identifies the primary content.

Navigation

Identifies major groups of links used for navigating the app.

Region

Identifies significant areas. Usually set with AccessState::Label.

Search

Identifies the search area or form.

Alert

Identifies important, and usually time-sensitive, information.

Log

Identifies a widget that creates a live region where new information is added in a meaningful order and old information may disappear.

Marquee

Identifies a live region containing non-essential information which changes frequently.

Status

Identifies a live region containing advisory information for the user that is not important enough to be an alert.

Timer

Indicates to assistive technologies that a widget is a numerical counter listing the amount of elapsed time from a starting point or the remaining time until an end point. Assistive technologies will not announce updates to a timer.

AlertDialog

Identifies a modal alert dialogs that interrupt a user’s workflow to communicate an important message and require a response.

Dialog

Identifies a widget that has content separate from the normal window and is presented as an overlay.