zng_view_api::access

Enum AccessState

#[non_exhaustive]
pub enum AccessState {

    AutoComplete(AutoComplete),
    Checked(Option<bool>),
    Current(CurrentKind),
    Disabled,
    ErrorMessage(AccessNodeId),
    Expanded(bool),
    Popup(Popup),
    Invalid(Invalid),
    Label(Txt),
    Level(NonZeroU32),
    Modal,
    MultiSelectable,
    Orientation(Orientation),
    Placeholder(Txt),
    ReadOnly,
    Required,
    Selected,
    Sort(SortDirection),
    ValueMax(f64),
    ValueMin(f64),
    Value(f64),
    ValueText(Txt),
    Live {
        indicator: LiveIndicator,
        atomic: bool,
        busy: bool,
    },
    ActiveDescendant(AccessNodeId),
    ColCount(usize),
    ColIndex(usize),
    ColSpan(usize),
    Controls(Vec<AccessNodeId>),
    DescribedBy(Vec<AccessNodeId>),
    Details(Vec<AccessNodeId>),
    FlowTo(Vec<AccessNodeId>),
    LabelledBy(Vec<AccessNodeId>),
    LabelledByChild,
    Owns(Vec<AccessNodeId>),
    ItemIndex(usize),
    RowCount(usize),
    RowIndex(usize),
    RowSpan(usize),
    ItemCount(usize),
    Lang(LanguageIdentifier),
    ScrollHorizontal(f32),
    ScrollVertical(f32),

}

Accessibility attribute of a node in the accessibility tree.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AutoComplete(AutoComplete)

Inputting text triggers display of one or more predictions of the user’s intended value for a ComboBox, SearchBox, or TextInput.

Checked(Option<bool>)

If the widget is checked (Some(true)), unchecked (Some(false)), or if the checked status is indeterminate (None).

Current(CurrentKind)

Indicates that the widget represents a current item.

Disabled

Indicates that the widget is perceivable but disabled, so it is not editable or otherwise operable.

ErrorMessage(AccessNodeId)

Indicates that the widget is an error message for the referenced node.

The other widget must be Invalid.

Expanded(bool)

Indicate that the widget toggles the visibility of related widgets.

Use Controls, or Owns to indicate the widgets that change visibility based on this value.

Popup(Popup)

Indicates the availability and type of interactive popup widget.

Invalid(Invalid)

Indicates the entered value does not conform to the format expected by the application.

Label(Txt)

Defines a string value that labels the widget.

Level(NonZeroU32)

Defines the hierarchical level of a widget within a structure.

Modal

Indicates whether the widget is modal when displayed.

MultiSelectable

Indicates that the user may select more than one item from the current selectable descendants.

Orientation(Orientation)

Indicates whether the widget’s orientation is horizontal, vertical, or unknown/ambiguous.

Placeholder(Txt)

Short hint (a word or short phrase) intended to help the user with data entry when a form control has no value.

ReadOnly

Indicates that the widget is not editable, but is otherwise operable.

Required

Indicates that user input is required on the widget before a form may be submitted.

Selected

Indicates that the widget is selected.

Sort(SortDirection)

Indicates if items in a table or grid are sorted in ascending or descending order.

ValueMax(f64)

Defines the maximum value (inclusive).

ValueMin(f64)

Defines the minimum value (inclusive).

Value(f64)

Defines the current value.

ValueText(Txt)

Defines a human readable version of the Value.

Live

Indicate that the widget can change.

Fields

indicator: LiveIndicator

How the changes must be notified.

atomic: bool

If the live region must be re-read entirely after each update.

busy: bool

Indicates the live area being modified and that assistive technologies may want to wait until the changes are complete before informing the user about the update.

ActiveDescendant(AccessNodeId)

Identifies the currently active widget when focus is on a composite widget, ComboBox, TextInput, Group, or Application.

ColCount(usize)

Defines the total number of columns in a Table, Grid, or TreeGrid when not all columns are present in tree.

The value 0 indicates that not all columns are in the widget and the application cannot determinate the exact number.

ColIndex(usize)

Defines a widget’s column index in the parent table or grid.

ColSpan(usize)

Defines the number of columns spanned by the widget in the parent table or grid.

Controls(Vec<AccessNodeId>)

Identifies the widget(s) whose contents or presence are controlled by this widget.

DescribedBy(Vec<AccessNodeId>)

Identifies the widget(s) that describes this widget.

Details(Vec<AccessNodeId>)

Identifies the widget(s) that provide additional information related to this widget.

FlowTo(Vec<AccessNodeId>)

Options for next widget to read.

LabelledBy(Vec<AccessNodeId>)

Identifies the widget(s) that labels the widget it is applied to.

LabelledByChild

Uses the widget children as LabelledBy.

Owns(Vec<AccessNodeId>)

Identifies widget(s) in order to define a visual, functional, or contextual relationship between a parent and its child widgets when the tree hierarchy cannot be used to represent the relationship.

ItemIndex(usize)

Defines the widget’s number or position in the current set of list items or tree items when not all items are present in the tree.

RowCount(usize)

Defines the total number of rows in a Table, Grid, or TreeGrid when not all rows are present in tree.

The value 0 indicates that not all rows are in the widget and the application cannot determinate the exact number.

RowIndex(usize)

Defines a widget’s row index in the parent table or grid.

RowSpan(usize)

Defines the number of rows spanned by the widget in the parent table or grid.

ItemCount(usize)

Defines the number of items in the current set of list items or tree items when not all items in the set are present in the tree.

Lang(LanguageIdentifier)

Language of texts inside the widget and descendants.

ScrollHorizontal(f32)

Normalized (0..1) horizontal scroll, 0 is showing the content leftmost edge, 1 is showing the content the rightmost edge.

ScrollVertical(f32)

Normalized (0..1) vertical scroll, 0 is showing the content topmost edge, 1 is showing the content the bottommost edge.