zng_view_api

Module audio

Audio device types.

Structs

AudioDecoded

Represents a partial or fully decoded audio.

AudioDeviceCapability

Represents audio device input/output capabilities.

AudioDeviceId

Audio device ID.

AudioDeviceInfo

Info about an input or output device.

AudioEncodeId

Id of an audio encode task.

AudioEncodeRequest

Represent a image encode request.

AudioFormat

Represents an audio codec capability.

AudioFormatCapability

Capabilities of an AudioFormat implementation.

AudioId

Id of a decoded or on demand decoding audio track in the cache.

AudioMetadata

Represents decoded header metadata about an audio track.

AudioMix

Represents an audio source.

AudioOutputConfig

Audio playback config.

AudioOutputId

Audio playback stream ID.

AudioOutputOpenData

Represents an audio output stream capabilities.

AudioOutputRequest

Represents a connection request to an audio output device.

AudioOutputUpdateRequest

Represents an audio playback update request.

AudioPlayId

Audio playback request ID.

AudioPlayRequest

Represents an audio playback request.

AudioRequest

Represent an audio load/decode request.

AudioStreamMode

Represents steam capability of an audio device.

AudioTrackMetadata

Represents decoded header metadata about a track position in the container represented by another audio.

AudioTracksMode

Defines what tracks are decoded from multi track containers.

SampleRate

Represents the minimum and maximum sample rate per audio channel.

Enums

AudioDataFormat

Format of the audio bytes.

AudioMixLayer

Represents an audio source component.

AudioOutputState

Represents the playback state if an audio output stream.

BufferSize

Represents the minimum and maximum supported buffer size for the device.