Enum zng_view_api::display_list::DisplayItem

pub enum DisplayItem {

    Reuse {
        frame_id: FrameId,
        seg_id: SegmentId,
        start: usize,
        end: usize,
    },
    PushReferenceFrame {
        id: ReferenceFrameId,
        transform: FrameValue<PxTransform>,
        transform_style: TransformStyle,
        is_2d_scale_translation: bool,
    },
    PopReferenceFrame,
    PushStackingContext {
        transform_style: TransformStyle,
        blend_mode: MixBlendMode,
        filters: Box<[FilterOp]>,
    },
    PopStackingContext,
    PushClipRect {
        clip_rect: PxRect,
        clip_out: bool,
    },
    PushClipRoundedRect {
        clip_rect: PxRect,
        corners: PxCornerRadius,
        clip_out: bool,
    },
    PopClip,
    PushMask {
        image_id: ImageTextureId,
        rect: PxRect,
    },
    PopMask,
    Border {
        bounds: PxRect,
        widths: PxSideOffsets,
        sides: [BorderSide; 4],
        radius: PxCornerRadius,
    },
    NinePatchBorder {
        bounds: PxRect,
        source: NinePatchSource,
        widths: PxSideOffsets,
        fill: bool,
        repeat_horizontal: RepeatMode,
        repeat_vertical: RepeatMode,
    },
    Text {
        clip_rect: PxRect,
        font_id: FontId,
        glyphs: Box<[GlyphInstance]>,
        color: FrameValue<Rgba>,
        options: GlyphOptions,
    },
    Image {
        clip_rect: PxRect,
        image_id: ImageTextureId,
        image_size: PxSize,
        rendering: ImageRendering,
        alpha_type: AlphaType,
        tile_size: PxSize,
        tile_spacing: PxSize,
    },
    Color {
        clip_rect: PxRect,
        color: FrameValue<Rgba>,
    },
    BackdropFilter {
        clip_rect: PxRect,
        filters: Box<[FilterOp]>,
    },
    LinearGradient {
        clip_rect: PxRect,
        start_point: Point2D<f32, Px>,
        end_point: Point2D<f32, Px>,
        extend_mode: ExtendMode,
        stops: Box<[GradientStop]>,
        tile_origin: PxPoint,
        tile_size: PxSize,
        tile_spacing: PxSize,
    },
    RadialGradient {
        clip_rect: PxRect,
        center: Point2D<f32, Px>,
        radius: Size2D<f32, Px>,
        start_offset: f32,
        end_offset: f32,
        extend_mode: ExtendMode,
        stops: Box<[GradientStop]>,
        tile_origin: PxPoint,
        tile_size: PxSize,
        tile_spacing: PxSize,
    },
    ConicGradient {
        clip_rect: PxRect,
        center: Point2D<f32, Px>,
        angle: AngleRadian,
        start_offset: f32,
        end_offset: f32,
        extend_mode: ExtendMode,
        stops: Box<[GradientStop]>,
        tile_origin: PxPoint,
        tile_size: PxSize,
        tile_spacing: PxSize,
    },
    Line {
        clip_rect: PxRect,
        color: Rgba,
        style: LineStyle,
        orientation: LineOrientation,
    },
    PushExtension {
        extension_id: ApiExtensionId,
        payload: ApiExtensionPayload,
    },
    PopExtension {
        extension_id: ApiExtensionId,
    },
    SetBackfaceVisibility {
        visible: bool,
    },

}

Display item in a DisplayList.

Variants

Reuse

Fields

frame_id: FrameId

seg_id: SegmentId

start: usize

end: usize

PushReferenceFrame

Fields

id: ReferenceFrameId

transform: FrameValue<PxTransform>

transform_style: TransformStyle

is_2d_scale_translation: bool

PopReferenceFrame

PushStackingContext

Fields

transform_style: TransformStyle

blend_mode: MixBlendMode

filters: Box<[FilterOp]>

PopStackingContext

PushClipRect

Fields

clip_rect: PxRect

clip_out: bool

PushClipRoundedRect

Fields

clip_rect: PxRect

corners: PxCornerRadius

clip_out: bool

PopClip

PushMask

Fields

image_id: ImageTextureId

rect: PxRect

PopMask

Border

Fields

bounds: PxRect

widths: PxSideOffsets

sides: [BorderSide; 4]

radius: PxCornerRadius

NinePatchBorder

Fields

bounds: PxRect

source: NinePatchSource

widths: PxSideOffsets

fill: bool

repeat_horizontal: RepeatMode

repeat_vertical: RepeatMode

Text

Fields

clip_rect: PxRect

font_id: FontId

glyphs: Box<[GlyphInstance]>

color: FrameValue<Rgba>

options: GlyphOptions

Image

Fields

clip_rect: PxRect

image_id: ImageTextureId

image_size: PxSize

rendering: ImageRendering

alpha_type: AlphaType

tile_size: PxSize

tile_spacing: PxSize

Color

Fields

clip_rect: PxRect

color: FrameValue<Rgba>

BackdropFilter

Fields

clip_rect: PxRect

filters: Box<[FilterOp]>

LinearGradient

Fields

clip_rect: PxRect

start_point: Point2D<f32, Px>

end_point: Point2D<f32, Px>

extend_mode: ExtendMode

stops: Box<[GradientStop]>

tile_origin: PxPoint

tile_size: PxSize

tile_spacing: PxSize

RadialGradient

Fields

clip_rect: PxRect

center: Point2D<f32, Px>

radius: Size2D<f32, Px>

start_offset: f32

end_offset: f32

extend_mode: ExtendMode

stops: Box<[GradientStop]>

tile_origin: PxPoint

tile_size: PxSize

tile_spacing: PxSize

ConicGradient

Fields

clip_rect: PxRect

center: Point2D<f32, Px>

angle: AngleRadian

start_offset: f32

end_offset: f32

extend_mode: ExtendMode

stops: Box<[GradientStop]>

tile_origin: PxPoint

tile_size: PxSize

tile_spacing: PxSize

Line

Fields

clip_rect: PxRect

color: Rgba

style: LineStyle

orientation: LineOrientation

PushExtension

Fields

extension_id: ApiExtensionId

payload: ApiExtensionPayload

PopExtension

Fields

extension_id: ApiExtensionId

SetBackfaceVisibility

Fields

visible: bool

Implementations

impl DisplayItem

pub fn update_transform(&mut self, t: &FrameValueUpdate<PxTransform>) -> bool

Update the value and returns if a new full frame must be rendered.

pub fn update_float(&mut self, t: &FrameValueUpdate<f32>) -> bool

Update the value and returns if a new full frame must be rendered.

pub fn update_color(&mut self, t: &FrameValueUpdate<Rgba>) -> bool

Update the value and returns if a new full frame must be rendered.