zng_view_api

Module image

Image types.

Structs

ColorType

Basic info about a color model.

ImageDecoded

Represents a partial or fully decoded image.

ImageEncodeId

Id of an image encode task.

ImageEncodeRequest

Represent a image encode request.

ImageEntriesMode

Defines what images are decoded from multi image containers.

ImageEntryMetadata

Represents decoded header metadata about an image position in a container represented by another image.

ImageFormat

Represents an image codec capability.

ImageFormatCapability

Capabilities of an ImageFormat implementation.

ImageId

Id of a decoded image in the cache.

ImageMetadata

Represents decoded header metadata about an image.

ImageRequest

Represent a image load/decode request.

ImageTextureId

Id of an image loaded in a renderer.

Enums

ImageDataFormat

Format of the image bytes.

ImageDownscaleMode

Defines how an image is downscaled after decoding.

ImageEntryKind

Kind of image container entry an image was decoded from.

ImageMaskMode

Defines how the A8 image mask pixels are to be derived from a source mask image.

PartialImageKind

Represents what kind of partial data was decoded.