zng_view_api::window

Enum CursorIcon

#[non_exhaustive]
pub enum CursorIcon {

    Default,
    ContextMenu,
    Help,
    Pointer,
    Progress,
    Wait,
    Cell,
    Crosshair,
    Text,
    VerticalText,
    Alias,
    Copy,
    Move,
    NoDrop,
    NotAllowed,
    Grab,
    Grabbing,
    EResize,
    NResize,
    NeResize,
    NwResize,
    SResize,
    SeResize,
    SwResize,
    WResize,
    EwResize,
    NsResize,
    NeswResize,
    NwseResize,
    ColResize,
    RowResize,
    AllScroll,
    ZoomIn,
    ZoomOut,

}

Named system dependent cursor icon.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Default

The platform-dependent default cursor. Often rendered as arrow.

ContextMenu

A context menu is available for the object under the cursor. Often rendered as an arrow with a small menu-like graphic next to it.

Help

Help is available for the object under the cursor. Often rendered as a question mark or a balloon.

Pointer

The cursor is a pointer that indicates a link. Often rendered as the backside of a hand with the index finger extended.

Progress

A progress indicator. The program is performing some processing, but is different from CursorIcon::Wait in that the user may still interact with the program.

Wait

Indicates that the program is busy and the user should wait. Often rendered as a watch or hourglass.

Cell

Indicates that a cell or set of cells may be selected. Often rendered as a thick plus-sign with a dot in the middle.

Crosshair

A simple crosshair (e.g., short line segments resembling a “+” sign). Often used to indicate a two dimensional bitmap selection mode.

Text

Indicates text that may be selected. Often rendered as an I-beam.

VerticalText

Indicates vertical-text that may be selected. Often rendered as a horizontal I-beam.

Alias

Indicates an alias of/shortcut to something is to be created. Often rendered as an arrow with a small curved arrow next to it.

Copy

Indicates something is to be copied. Often rendered as an arrow with a small plus sign next to it.

Move

Indicates something is to be moved.

NoDrop

Indicates that the dragged item cannot be dropped at the current cursor location. Often rendered as a hand or pointer with a small circle with a line through it.

NotAllowed

Indicates that the requested action will not be carried out. Often rendered as a circle with a line through it.

Grab

Indicates that something can be grabbed (dragged to be moved). Often rendered as the backside of an open hand.

Grabbing

Indicates that something is being grabbed (dragged to be moved). Often rendered as the backside of a hand with fingers closed mostly out of view.

EResize

The east border to be moved.

NResize

The north border to be moved.

NeResize

The north-east corner to be moved.

NwResize

The north-west corner to be moved.

SResize

The south border to be moved.

SeResize

The south-east corner to be moved.

SwResize

The south-west corner to be moved.

WResize

The west border to be moved.

EwResize

The east and west borders to be moved.

NsResize

The south and north borders to be moved.

NeswResize

The north-east and south-west corners to be moved.

NwseResize

The north-west and south-east corners to be moved.

ColResize

Indicates that the item/column can be resized horizontally. Often rendered as arrows pointing left and right with a vertical bar separating them.

RowResize

Indicates that the item/row can be resized vertically. Often rendered as arrows pointing up and down with a horizontal bar separating them.

AllScroll

Indicates that the something can be scrolled in any direction. Often rendered as arrows pointing up, down, left, and right with a dot in the middle.

ZoomIn

Indicates that something can be zoomed in. Often rendered as a magnifying glass with a “+” in the center of the glass.

ZoomOut

Indicates that something can be zoomed in. Often rendered as a magnifying glass with a “-” in the center of the glass.

Implementations

impl CursorIcon

pub const ALL: &'static [CursorIcon]

All cursor icons.

pub fn size_and_spot(&self, scale_factor: Factor) -> (PxSize, PxPoint)

Estimated icon size and click spot in that size.