Module zng_view_api::access

Accessibility and automation types.

Structs

AccessNode

Represents a widget in the access info tree.

AccessNodeId

Identifies an accessibility widget node.

AccessNodeRef

Reference an access node in a tree.

AccessTree

Accessibility info tree for a window.

AccessTreeBuilder

Accessibility info tree builder.

AccessTreeUpdate

Update for accessibility info tree for a window.

AutoComplete

Defines how inputting text could trigger display of one or more predictions of the user’s intended value.

Invalid

Defines the kind of invalid data error of a widget.

Enums

AccessCmd

Accessibility command.

AccessCmdName

Accessibility command without associated data.

AccessRole

Accessibility role of a node in the accessibility tree.

AccessState

Accessibility attribute of a node in the accessibility tree.

CurrentKind

Kind of current item a widget represents.

LiveIndicator

Defines how a live update is communicated to the user.

Orientation

Widget orientation.

Popup

Popup type.

ScrollCmd

Accessibility scroll command.

SortDirection

Sort direction.