Module zng_view_api::display_list

Frame builder types.

Structs

DisplayList

Represents a finalized display list.

DisplayListBuilder

Represents a builder for display items that will be rendered in the view process.

FrameValueId

Frame value binding ID.

FrameValueUpdate

Represents an update targeting a previously setup FrameValue.

ReuseRange

Represents a display list reuse range.

ReuseStart

Represents the start of a display list reuse range.

Enums

DisplayItem

Display item in a DisplayList.

FilterOp

Represents one of the filters applied to a stacking context.

FrameValue

Represents a frame value that may be updated.

NinePatchSource

Nine-patch image source.

Type Aliases

SegmentId

Id of the offset a parallel list folded onto the main list is at.