Enum zng_view_api::Event

pub enum Event {

    Inited(Inited),
    Disconnected(ViewProcessGen),
    WindowOpened(WindowId, WindowOpenData),
    HeadlessOpened(WindowId, HeadlessOpenData),
    WindowOrHeadlessOpenError {
        id: WindowId,
        error: Txt,
    },
    FrameRendered(EventFrameRendered),
    WindowChanged(WindowChanged),
    DroppedFile {
        window: WindowId,
        file: PathBuf,
    },
    HoveredFile {
        window: WindowId,
        file: PathBuf,
    },
    HoveredFileCancelled(WindowId),
    FocusChanged {
        prev: Option<WindowId>,
        new: Option<WindowId>,
    },
    KeyboardInput {
        window: WindowId,
        device: DeviceId,
        key_code: KeyCode,
        state: KeyState,
        key: Key,
        key_modified: Key,
        text: Txt,
    },
    Ime {
        window: WindowId,
        ime: Ime,
    },
    MouseMoved {
        window: WindowId,
        device: DeviceId,
        coalesced_pos: Vec<DipPoint>,
        position: DipPoint,
    },
    MouseEntered {
        window: WindowId,
        device: DeviceId,
    },
    MouseLeft {
        window: WindowId,
        device: DeviceId,
    },
    MouseWheel {
        window: WindowId,
        device: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
    },
    MouseInput {
        window: WindowId,
        device: DeviceId,
        state: ButtonState,
        button: MouseButton,
    },
    TouchpadPressure {
        window: WindowId,
        device: DeviceId,
        pressure: f32,
        stage: i64,
    },
    AxisMotion {
        window: WindowId,
        device: DeviceId,
        axis: AxisId,
        value: f64,
    },
    Touch {
        window: WindowId,
        device: DeviceId,
        touches: Vec<TouchUpdate>,
    },
    ScaleFactorChanged {
        monitor: MonitorId,
        windows: Vec<WindowId>,
        scale_factor: f32,
    },
    MonitorsChanged(Vec<(MonitorId, MonitorInfo)>),
    ColorSchemeChanged(WindowId, ColorScheme),
    WindowCloseRequested(WindowId),
    WindowClosed(WindowId),
    ImageMetadataLoaded {
        image: ImageId,
        size: PxSize,
        ppi: Option<ImagePpi>,
        is_mask: bool,
    },
    ImageLoaded(ImageLoadedData),
    ImagePartiallyLoaded {
        image: ImageId,
        partial_size: PxSize,
        ppi: Option<ImagePpi>,
        is_opaque: bool,
        is_mask: bool,
        partial_pixels: IpcBytes,
    },
    ImageLoadError {
        image: ImageId,
        error: Txt,
    },
    ImageEncoded {
        image: ImageId,
        format: Txt,
        data: IpcBytes,
    },
    ImageEncodeError {
        image: ImageId,
        format: Txt,
        error: Txt,
    },
    FrameImageReady {
        window: WindowId,
        frame: FrameId,
        image: ImageId,
        selection: PxRect,
    },
    FontsChanged,
    FontAaChanged(FontAntiAliasing),
    MultiClickConfigChanged(MultiClickConfig),
    AnimationsConfigChanged(AnimationsConfig),
    KeyRepeatConfigChanged(KeyRepeatConfig),
    TouchConfigChanged(TouchConfig),
    LocaleChanged(LocaleConfig),
    DeviceAdded(DeviceId),
    DeviceRemoved(DeviceId),
    DeviceMouseMotion {
        device: DeviceId,
        delta: Vector2D<f64, ()>,
    },
    DeviceMouseWheel {
        device: DeviceId,
        delta: MouseScrollDelta,
    },
    DeviceMotion {
        device: DeviceId,
        axis: AxisId,
        value: f64,
    },
    DeviceButton {
        device: DeviceId,
        button: ButtonId,
        state: ButtonState,
    },
    DeviceKey {
        device: DeviceId,
        key_code: KeyCode,
        state: KeyState,
    },
    MsgDialogResponse(DialogId, MsgDialogResponse),
    FileDialogResponse(DialogId, FileDialogResponse),
    AccessInit {
        window: WindowId,
    },
    AccessCommand {
        window: WindowId,
        target: AccessNodeId,
        command: AccessCmd,
    },
    LowMemory,
    RecoveredFromComponentPanic {
        component: Txt,
        recover: Txt,
        panic: Txt,
    },
    ExtensionEvent(ApiExtensionId, ApiExtensionPayload),

}

System and User events sent from the View Process.

Variants

Inited(Inited)

View-process inited.

Disconnected(ViewProcessGen)

The event channel disconnected, probably because the view-process crashed.

The ViewProcessGen is the generation of the view-process that was lost, it must be passed to Controller::handle_disconnect.

WindowOpened(WindowId, WindowOpenData)

Window, context and renderer have finished initializing and is ready to receive commands.

HeadlessOpened(WindowId, HeadlessOpenData)

Headless context and renderer have finished initializing and is ready to receive commands.

WindowOrHeadlessOpenError

Fields

id: WindowId

Id from the request.

error: Txt

Error message.

Window open or headless context open request failed.

FrameRendered(EventFrameRendered)

A frame finished rendering.

EventsCleared is not send after this event.

WindowChanged(WindowChanged)

Window moved, resized, or minimized/maximized etc.

This event aggregates events moves, resizes and other state changes into a single event to simplify tracking composite changes, for example, the window changes size and position when maximized, this can be trivially observed with this event.

The EventCause can be used to identify a state change initiated by the app.

DroppedFile

Fields

window: WindowId

Window that received the file drop.

file: PathBuf

Path to the file that was dropped.

A file has been dropped into the window.

When the user drops multiple files at once, this event will be emitted for each file separately.

HoveredFile

Fields

window: WindowId

Window that was hovered by drag-drop.

file: PathBuf

Path to the file being dragged.

A file is being hovered over the window.

When the user hovers multiple files at once, this event will be emitted for each file separately.

HoveredFileCancelled(WindowId)

A file was hovered, but has exited the window.

There will be a single event triggered even if multiple files were hovered.

FocusChanged

Fields

prev: Option<WindowId>

Window that lost focus.

new: Option<WindowId>

Window that got focus.

App window(s) focus changed.

KeyboardInput

Fields

window: WindowId

Window that received the key event.

device: DeviceId

Device that generated the key event.

key_code: KeyCode

Physical key.

state: KeyState

If the key was pressed or released.

key: Key

Semantic key.

Pressing Shift+A key will produce Key::Char('a') in QWERTY keyboards, the modifiers are not applied.

key_modified: Key

Semantic key modified by the current active modifiers.

Pressing Shift+A key will produce Key::Char('A') in QWERTY keyboards, the modifiers are applied.

text: Txt

Text typed.

This is only set during KeyState::Pressed of a key that generates text.

This is usually the key_modified char, but is also '\r' for Key::Enter. On Windows when a dead key was pressed earlier but cannot be combined with the character from this key press, the produced text will consist of two characters: the dead-key-character followed by the character resulting from this key press.

An event from the keyboard has been received.

This event is only send if the window is focused, all pressed keys should be considered released after FocusChanged to None. Modifier keys receive special treatment, after they are pressed, the modifier key state is monitored directly so that the Released event is always send, unless the focus changed to none.

Ime

Fields

window: WindowId

Window that received the IME event.

ime: Ime

IME event.

IME composition event.

MouseMoved

Fields

window: WindowId

Window that received the cursor move.

device: DeviceId

Device that generated the cursor move.

coalesced_pos: Vec<DipPoint>

Cursor positions in between the previous event and this one.

position: DipPoint

Cursor position, relative to the window top-left in device independent pixels.

The mouse cursor has moved on the window.

This event can be coalesced, i.e. multiple cursor moves packed into the same event.

MouseEntered

Fields

window: WindowId

Window that now is hovered by the cursor.

device: DeviceId

Device that generated the cursor move event.

The mouse cursor has entered the window.

MouseLeft

Fields

window: WindowId

Window that is no longer hovered by the cursor.

device: DeviceId

Device that generated the cursor move event.

The mouse cursor has left the window.

MouseWheel

Fields

window: WindowId

Window that was hovered by the cursor when the mouse wheel was used.

device: DeviceId

Device that generated the mouse wheel event.

delta: MouseScrollDelta

Delta of change in the mouse scroll wheel state.

phase: TouchPhase

Touch state if the device that generated the event is a touchpad.

A mouse wheel movement or touchpad scroll occurred.

MouseInput

Fields

window: WindowId

Window that was hovered by the cursor when the mouse button was used.

device: DeviceId

Mouse device that generated the event.

state: ButtonState

If the button was pressed or released.

button: MouseButton

The mouse button.

An mouse button press has been received.

TouchpadPressure

Fields

window: WindowId

Window that was hovered when the touchpad was touched.

device: DeviceId

Touchpad device.

pressure: f32

Pressure level between 0 and 1.

stage: i64

Click level.

Touchpad pressure event.

AxisMotion

Fields

window: WindowId

Window that was focused when the motion was realized.

device: DeviceId

Analog device.

axis: AxisId

Axis.

value: f64

Motion value.

Motion on some analog axis. May report data redundant to other, more specific events.

Touch

Fields

window: WindowId

Window that was touched.

device: DeviceId

Touch device.

touches: Vec<TouchUpdate>

Coalesced touch updates, never empty.

Touch event has been received.

ScaleFactorChanged

Fields

monitor: MonitorId

Monitor that has changed.

windows: Vec<WindowId>

Windows affected by this change.

Note that a window’s scale factor can also change if it is moved to another monitor, the Event::WindowChanged event notifies this using the WindowChanged::monitor.

scale_factor: f32

The new scale factor.

The monitor’s scale factor has changed.

MonitorsChanged(Vec<(MonitorId, MonitorInfo)>)

The available monitors have changed.

ColorSchemeChanged(WindowId, ColorScheme)

The preferred color scheme for a window has changed.

WindowCloseRequested(WindowId)

The window has been requested to close.

WindowClosed(WindowId)

The window has closed.

ImageMetadataLoaded

Fields

image: ImageId

The image that started loading.

size: PxSize

The image pixel size.

ppi: Option<ImagePpi>

The image pixels-per-inch metadata.

is_mask: bool

The image is a single channel R8.

An image resource already decoded size and PPI.

ImageLoaded(ImageLoadedData)

An image resource finished decoding.

ImagePartiallyLoaded

Fields

image: ImageId

The image that has decoded more pixels.

partial_size: PxSize

The size of the decoded pixels, can be different then the image size if the image is not interlaced.

ppi: Option<ImagePpi>

The image pixels-per-inch metadata.

is_opaque: bool

If the decoded pixels so-far are all opaque (255 alpha).

is_mask: bool

If the decoded pixels so-far are a single channel.

partial_pixels: IpcBytes

Updated BGRA8 pre-multiplied pixel buffer or R8 if is_mask. This includes all the pixels decoded so-far.

An image resource, progressively decoded has decoded more bytes.

ImageLoadError

Fields

image: ImageId

The image that failed to decode.

error: Txt

The error message.

An image resource failed to decode, the image ID is not valid.

ImageEncoded

Fields

image: ImageId

The image that finished encoding.

format: Txt

The format of the encoded data.

data: IpcBytes

The encoded image data.

An image finished encoding.

ImageEncodeError

Fields

image: ImageId

The image that failed to encode.

format: Txt

The encoded format that was requested.

error: Txt

The error message.

An image failed to encode.

FrameImageReady

Fields

window: WindowId

Window that had pixels copied.

frame: FrameId

The frame that was rendered when the pixels where copied.

image: ImageId

The frame image.

selection: PxRect

The pixel selection relative to the top-left.

An image generated from a rendered frame is ready.

FontsChanged

System fonts have changed.

FontAaChanged(FontAntiAliasing)

System text anti-aliasing configuration has changed.

MultiClickConfigChanged(MultiClickConfig)

System double-click definition changed.

AnimationsConfigChanged(AnimationsConfig)

System animations config changed.

KeyRepeatConfigChanged(KeyRepeatConfig)

System definition of pressed key repeat event changed.

TouchConfigChanged(TouchConfig)

System touch config changed.

LocaleChanged(LocaleConfig)

System locale changed.

DeviceAdded(DeviceId)

Device added or installed.

DeviceRemoved(DeviceId)

Device removed.

DeviceMouseMotion

Fields

device: DeviceId

Device that generated the event.

delta: Vector2D<f64, ()>

Delta of change in the cursor position.

Mouse pointer motion.

The values if the delta of movement (x, y), not position.

DeviceMouseWheel

Fields

device: DeviceId

Mouse device that generated the event.

delta: MouseScrollDelta

Delta of change in the mouse scroll wheel state.

Mouse scroll wheel turn.

DeviceMotion

Fields

device: DeviceId

Device that generated the event.

axis: AxisId

Device dependent axis of the motion.

value: f64

Device dependent value.

Motion on some analog axis.

This includes the mouse device and any other that fits.

DeviceButton

Fields

device: DeviceId

Device that generated the event.

button: ButtonId

Device dependent button that was used.

state: ButtonState

If the button was pressed or released.

Device button press or release.

DeviceKey

Fields

device: DeviceId

Device that generated the key event.

key_code: KeyCode

Physical key.

state: KeyState

If the key was pressed or released.

Device key press or release.

MsgDialogResponse(DialogId, MsgDialogResponse)

User responded to a native message dialog.

FileDialogResponse(DialogId, FileDialogResponse)

User responded to a native file dialog.

AccessInit

Fields

window: WindowId

Window that received the first accessibility request.

Accessibility info tree was requested for the first time on a window.

AccessCommand

Fields

window: WindowId

Window that had pixels copied.

target: AccessNodeId

Target widget.

command: AccessCmd

Command.

Accessibility command.

LowMemory

System low memory warning, some platforms may kill the app if it does not release memory.

RecoveredFromComponentPanic

Fields

component: Txt

Component identifier.

recover: Txt

How the view-process recovered from the panic.

panic: Txt

The panic.

An internal component panicked, but the view-process manager to recover from it without needing to respawn.

ExtensionEvent(ApiExtensionId, ApiExtensionPayload)

Represents a custom event send by the extension.

Implementations

impl Event

pub fn coalesce(&mut self, other: Event) -> Result<(), Event>

Change self to incorporate other or returns other if both events cannot be coalesced.