Module zng_view_api::ipc

IPC types.

Structs

Disconnected

Channel disconnected error.

EventSender

Event channel entry-point.

IpcBytes

Immutable shared memory that can be send fast over IPC.

IpcBytesReceiver

Bytes receiver.

IpcBytesSender

Bytes sender.

RequestReceiver

Requests channel end-point.

ResponseSender

Responses channel entry-point.

ViewChannels

Channels that must be used for implementing a view-process.

Functions

bytes_channel

Create a bytes channel.

connect_view_process

Start the view-process server and waits for (request, response, event).