Module zng_view_api::window

Window, surface and frame types.

Structs

CursorImage

Defines a custom mouse cursor.

EventFrameRendered

Event::FrameRendered payload.

FrameId

Identifier of a frame or frame update.

FrameRequest

Data for rendering a new frame.

FrameUpdateRequest

Data for rendering a new frame that is derived from the current frame.

FrameWaitId

Identifies a frame request for collaborative resize in WindowChanged.

HeadlessOpenData

Information about a successfully opened headless surface.

HeadlessRequest

Configuration of a new headless surface.

MonitorId

Monitor screen ID in channel.

MonitorInfo

Information about a monitor screen.

VideoMode

Exclusive video mode info.

WindowChanged

Event::WindowChanged payload.

WindowId

Window ID in channel.

WindowOpenData

Information about a successfully opened window.

WindowRequest

Configuration of a new window.

WindowStateAll

Represents the properties of a window that affect its position, size and state.

Enums

CursorIcon

Describes the appearance of the mouse cursor.

EventCause

Cause of a window state change.

FocusIndicator

Represents a focus request indicator.

FrameCapture

Frame image capture request.

RenderMode

Render backend preference.

WindowState

Window state.